Move painting colors tests to material
