+++ dev/bots/check_tests_cross_imports.dart
@@ -154,7 +154,6 @@ class TestsCrossImportChecker {
     'packages/flutter/test/painting/system_fonts_test.dart',
     'packages/flutter/test/painting/decoration_image_lerp_test.dart',
     'packages/flutter/test/painting/continuous_rectangle_border_test.dart',
-    'packages/flutter/test/painting/colors_test.dart',
   };
   static final Set<String> knownPhysicsCrossImports = <String>{};
   static final Set<String> knownRenderingCrossImports = <String>{
